+++ trashed-by.php
@@ -2,9 +2,9 @@
 /**
  * Plugin Name: Trashed By
  * Version:     1.3
- * Plugin URI:  http://coffee2code.com/wp-plugins/trashed-by/
+ * Plugin URI:  https://coffee2code.com/wp-plugins/trashed-by/
  * Author:      Scott Reilly
- * Author URI:  http://coffee2code.com/
+ * Author URI:  https://coffee2code.com/
  * Text Domain: trashed-by
  * License:     GPLv2 or later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html