+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _(in-progress)_
 * Change: Note compatibility through WP 5.4+
+* Change: Update links to coffee2code.com to be HTTPS
 
 ## 1.3 _(2020-03-17)_
 * New: Add CHANGELOG.md and move all but most recent changelog entries into it