+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 # Changelog
 
-## _(in-progress)_
+## 1.3.1 _(2020-07-27)_
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS
 * Change: Fix typo in docs
 * Change: Remove a tag from readme.txt's 'Tags' field
-* Unit tests:
-    * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests (and delete commented-out code)
+* Unit tests: Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests (and delete commented-out code)
 
 ## 1.3 _(2020-03-17)_
 * New: Add CHANGELOG.md and move all but most recent changelog entries into it